added dev mode clerk
