+++ app/globals.css
@@ -5,3 +5,21 @@
   margin: 0;
 }
 
+.auth-button {
+  border-style: none;
+  background-color: transparent;
+  color: white;
+  font-size: large;
+  cursor: pointer;
+}
+
+.auth-button:hover {
+  scale: 110%;
+  transition: all 0.3s;
+}
+
+.user-pic {
+  object-fit: cover;
+  height: 40px;
+  width: 40px;
+}
